Refactor MeikyuKingdomBasic
